Move tests to __test__ to appropriate folder // add prod ws url
