+++ .env.example
@@ -1,1 +1,2 @@
-REACT_APP_WEBSOCKETURL=ws://localhost:3001
+REACT_APP_DEV_WEBSOCKETURL=ws://localhost:3001
+REACT_APP_PROD_WEBSOCKETURL=wss://randomstring.execute-api.us-west-2.amazonaws.com/dev
